pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whirr"
dynamic = ["version"]
description = "Local experiment orchestration. Queue jobs, track metrics, wake up to results."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Syntropy Systems" }
]
keywords = ["ml", "experiment", "tracking", "orchestration", "wandb", "mlflow"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
metrics = [
    "psutil>=5.9",
]
dashboard = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
]
server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
]
ablate = []
dev = [
    "pytest>=7.0",
    "pytest-timeout>=2.0",
    "psutil>=5.9",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "jinja2>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.25.0",
]

[project.scripts]
whirr = "whirr.cli.main:app"

[project.urls]
Homepage = "https://github.com/syntropy-systems-oss/whirr"
Repository = "https://github.com/syntropy-systems-oss/whirr"
Issues = "https://github.com/syntropy-systems-oss/whirr/issues"

[tool.hatch.version]
path = "src/whirr/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/whirr"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "E501",
    "PLC0415",
    "PLR2004",
    "S101",  # pytest requires assert, everything else enforced
    "S108",
]
"src/whirr/cli/**" = [
    "B008",
    "C901",
    "FBT001",
    "FBT002",
    "FBT003",
    "PLC0415",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",  # conflicts with ruff format
    "ISC001",  # conflicts with ruff format
]

[tool.ruff.lint.flake8-copyright]
notice-rgx = "(?i)Copyright\\s+\\(c\\)\\s+Syntropy\\s+Systems"
min-file-size = 1

[tool.basedpyright]
typeCheckingMode = "all"
reportDeprecated = false
reportCallInDefaultInitializer = false

[dependency-groups]
dev = [
    "basedpyright>=1.37.1",
    "ruff>=0.14.11",
]
